feat: usePatchTaskIsDoneの作成 #180
ロジックを分離するため
